% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_axes_pos.R
\name{calc_axes_pos}
\alias{calc_axes_pos}
\title{Calculate axes positions for corradar plots.}
\usage{
calc_axes_pos(data, vars = NULL, max_iteration = 1000, min_degrees = 3)
}
\arguments{
\item{data}{Data file.}

\item{vars}{Variables to calculate axis positions for.}

\item{max_iteration}{Maxmimum amount of iterations for algorithm to run.}

\item{min_degrees}{Minimum degrees between each variable for visual spacing.}
}
\value{
A named numeric vector of axis positions in radians, wrapped to the
interval [0, 2π).
}
\description{
Calculate axes positions for corradar plots.
}
\examples{
data <- iris

iris_corradar_pos <- calc_axes_pos(data)

iris_corradar_pos
}
